+++ src/splits.cpp
@@ -29,7 +29,11 @@ RawMatrix cpp_edge_to_splits(const IntegerMatrix edge,
                             "Contact maintainer for advice");                   // # nocov
   }
   if (nTip[0] < 1) {
-    Rcpp::stop("Tree must contain tips.");
+    if (nTip[0] == 0) {
+      return RawMatrix(0, 0);
+    } else {
+      Rcpp::stop("Tree must contain non-negative number of tips.");
+    }
   }
   
   const uintx n_edge = edge.rows();
@@ -122,11 +126,12 @@ LogicalVector duplicated_splits(const RawMatrix splits,
     n_spare = n_tip % BIN_SIZE,
     check_bins = n_bin - (n_spare == 1)
   ;
-  if (n_bin != splits.cols()) {
-    Rcpp::stop("`splits` tip number is mis-specified.");
-  }
   if (n_split == 0) {
     return Rcpp::LogicalVector(0);
+  }
+  // assert(n_split > 0);
+  if (n_bin != splits.cols()) {
+    Rcpp::stop("`splits` tip number is mis-specified.");
   }
   
   RawMatrix compare(n_split, check_bins);