+++ tests/testthat/test-parsers.R
@@ -37,7 +37,7 @@ test_that("Matrix converts to phyDat", {
   mat <- matrix(c(1,0,1,0,1,0,1,0,0,1,0,1,0,1,0,1,2,2,2,2,2,2,2,"?"),
                 nrow = 3, byrow = TRUE)
   rownames(mat) <- LETTERS[1:3]
-  expect_equal(mat, PhyDatToMatrix(MatrixToPhyDat(mat)))
+  expect_equal(mat, as.matrix(MatrixToPhyDat(mat)))
 })
 
 test_that(".PhyDatWithContrast() fails gracefully", {
@@ -160,6 +160,20 @@ test_that("ReadNotes() reads notes", {
   expect_equal(notes[[3]][[2]], setNames("Tax1-Char3.", "taxon_a"))
 })
 
+test_that("ReadNotes() handles absence of character-taxon notes", {
+  expect_equal(ReadNotes(system.file("extdata/tests/taxon-notes.nex",
+                                     package = "TreeTools")),
+               structure(list(), names = character(0)))
+})
+
+test_that("ReadNotes() handles misspecified encoding", {
+  expect_message(
+    expect_equal(ReadNotes(system.file("extdata/tests/encoding.nex",
+                                     package = "TreeTools"))[[1]][[2]],
+                 setNames("\u0080ncoding.", "Two")),
+    "trying latin1 .*encoding")
+})
+
 test_that("ReadCharacters() reads CHARSTATELABELS", {
   labels <- ReadCharacters(system.file("extdata/input/dataset.nex",
                                        package = "TreeTools"))
@@ -199,5 +213,18 @@ test_that("NewickTree() works", {
 })
 
 test_that("as_newick() fails gracefully", {
-  expect_error(as_newick(matrix(0L, 8192 * 2L, 2L)))
+  expect_equal(as_newick(matrix(0L, 0L, 2L)), ";")
+  expect_equal(as_newick(matrix(1:0, 1L, 2L)), "(0);")
+  expect_equal(as_newick(Postorder(BalancedTree(4)$edge) - 1L),
+               as_newick(BalancedTree(4)$edge - 1L))
+  expect_error(as_newick(matrix(0L, 8192 * 2L, 2L)),
+               "Too many nodes")
+  expect_error(as_newick(matrix(0L, 3, 3)),
+               "`edge` must have two columns")
+  expect_error(as_newick(matrix(c(4, 4, 4, 1:3), 3, 2)),
+               "`min.edge.` must be zero")
+  expect_error(as_newick(matrix(c(3, NA, 3, 0:2), 3, 2)),
+               "`edge`.* NA")
+  expect_error(as_newick(matrix(c(4, 4, 3, 0:2), 3, 2)),
+               "`edge` is malformed")
 })