+++ NEWS.md
@@ -1,3 +1,28 @@
+# TreeTools 1.9.0 #
+
+## New methods and functions
+
+- `ZeroTaxonTree()` creates a `phylo` object with no leaves.
+
+- `DropTip()` gains new methods `DropTip.list()` and `DropTip.NULL()`.
+
+- `as.matrix.phylo()` converts a tree to a matrix representation, allowing
+  a tree to be passed as a constraint to `ImposeConstraint()`.
+  
+- `as.matrix.Splits()` and `as.matrix.phyDat()` methods added as synonyms to
+  `as.logical.Splits()` and `PhyDatToMatrix()`.
+
+## Improvements
+
+- Handle `TipLabels(0)` and `BalancedTree(0)`.
+
+- Support zero-leaf trees in `as.Splits()` and `duplicated.Splits()`.
+
+- Support non-identical tip labels in `as.Splits()`.
+
+- Try Latin-1 encoding if `ReadCharacters()` family fail under UTF-8.
+
+
 # TreeTools 1.8.0 #
 
 ## New methods and functions